build query builder w/ operators, expressions, add/remove, example criteria
every expression is stored in state, and operators are either AND or OR but not both at once. still needs styling cleanup and to actually tranlate the query to SQL. not yet validating or handling different types of values i.e. dates and numbers, so far it's all strings. also haven't added any match types. l_operand and r_operand stand for left and right operand, to be used later in translating our query to SQL. expressions can be added and removed. tried disabling the inputs when a column isn't picked, but that is not foolproof yet. also added stats action and reducer, which currently do nothing.