make Stats timeframe selection work well, Select controlled, graphs prettier
Fixed up the timeframe selection on the Stats page so it now works much more efficiently and doesn't depend too heavily on state. Made the Select component better with both a 'selected' prop that makes it a controlled component, and 'size' prop which allows devs to set the width easier. added some cool helper functions to help do some math for us, and AppConstants.